Thin Replica: BlockId instead of Cursor
During the desing phase with DA we missed the Ledger offset. This is an
integral part for applications to refer to the Ledger. Meaning, we need to
translate this concept into our KVB implementation. The easiest
solution is to use block ids. They fit the minimum requirement of being a
unique, monotonically increasing reference into the blockchain. Each block can
have mutliple transactions and key-value pairs. Meaning, in order to
deterministically compute the Ledger offset we need to persisit the order of
transactions/kv-pairs within each block when streaming updates to the thin
replica client.