Add Jaeger to support request tracing through Concord
Jaeger provides a C++ implementation of OpenTracing. We can use this
to time different portions of our request handling time.

https://jaegertracing.io/
https://github.com/jaegertracing/jaeger-client-cpp

An example "apiReqeust" span has been added in ApiConnection, which
measures the total time between finishing reading the request of the
wire, and preparing to serialize the response. In future changes, this
span should be passed along the request processing path, so that child
spans and span logs can be added to it as necessary.

Interesting side notes:

1. json.hpp has been removed from the utils directory, because it is
   now included in the Prereqs image. Inclusion in the Prereqs image
   was required to build Jaeger. Having a different version in the
   Concord source tree is both unnecessary, and also causes Concord to
   segfault when attempting to read genesis.json.

2. A config parameter has been added in the node scope, `jaeger_agent`,
   which provides the host or IP, and port of a jaeger-agent process
   ready to accept trace data.

3. Some preparation has been done for starting a "jaeger-agent" docker
   container, to funnel traces out to a future wavefront proxy.