Thin Replica: ReadState/ReadStateHash
First implementation of ReadState and ReadStateHash. This change introduces the
KVB app filter. For now, it is used to filter for DAML data only but can be
extended easily. In addition, this change introduces a new storage key for
DAML. Calculating the hash is done by leveraging the std::hash implemenation
for strings.