Define and add Thin Replica Client Library/application API.
This commit adds a definition for the intended API between the Thin Replica Client Library and the application consuming it to thin-replica-client/lib/thin_replica_client.hpp; documentation of the interface is also included in the header file. Additionally, this commit makes the changes needed to keep the existing example Thin Replica Client application working similarly to how it already did after having changed the ThinReplicaClient interface. Additionally, this commit introduces a README file for the thin-replica-client directory. The README includes a very brief description of what the Thin Replica mechanism is (which I thought warranted to create this README), but the main reason I decided to create the README now rather than later is so that it can serve as documentation on how to build the example Thin Replica Client application and run it against the DAML Nano cluster.