VB-2179: propagate tracing context over BFT network
This change adds a `trace_context` field to the `ConcordRequest`
message. In that field is serialized the OpenTracing context, so that
spans measured on each replica can be reported as part of the original
request span.

The field is packed by the KVBClient, and is unpacked by the
ConcordCommandsHandler, such that all execution engines get the
transfer logic for free.