VB-2232: management endpoints throw 500 when authenticated.
TokenAuthenticationProvider was autowiring HttpServletRequest in the constructor.  This was so we can log the requested URL.
It turns out that when you are using an actuator endpoint, this doesn't work.  The workaround was to check if the request context is a servlet context, then get the URI from that.

Tested by hand.  Not sure exactly how to put in some automated testing, but will look at more.