Speed up DamlTests
This is probably the only time that a sleep helps to shorten the time of
execution. Anyhow, this change introduces a sleep to the daml test tool
service so that Hermes can grab its output. Previously, due to the service
doesn't specify a command, the container shut down immediately and Hermes
kept waiting for output until it timed-out and continued with the test.
With this change we prevent the timeout by giving Hermes what it wants :)