vb-2521: Refactor helen to use vmbc-base
Changes to helen that use the new vmbc-base package.

Mostly changes to remove exceptions that are now in base, and use the new roles.

updating pom and removing redundant exception

Had to modify the way VmbcRoles work.  Occasionally getting an error because of statics use.

Bumping version of vmbc-base to 1.0.1-SNAPSHOT

Need to copy a change into ErrorCode (one of the messages changed).

Also changed to Dockerfile to build vmbc-base in line.  One of these days, properly build this separately, and push to artificatory.