Implement basic ReadState functionality in the Thin Replica Client.
This commit adds some basic implementation to ThinReplicaClient::Subscribe in the Thin Replica Client Library to allow subscribe to stream in the initial state from a Thin Replica Server. This commit also introduces extremely basic logic to collect hashes from F servers other than the one sending the state such that the state will not be returned unless at least F + 1 servers are found to agree on the state (however, this extremely basic logic gives up quickly if it cannot find agreement easily and does not include remotely robust error handling for cases involving incomplete connectivity to the servers and/or disagreement among the servers). Additional notable changes included in this commit:
* Implementation of the RingBufferUpdateQueue that provides an UpdateQueue implementation for transferring updates between the ThinReplicaClient object and application code.
* Minor changes to the UpdateQueue interface in light of concerns realized during the implementation of RingBufferUpdateQueue.
* Changes to the Thin Replica Client example application to ensure it meaningfully uses the ReadState functionality added to ThinReplicaClient::Subscribe.
* Changes to thin_replica_client/README.md to document the new expected behavior of the example application.