Regen pregen'd concord configs, to fix compatibility post-853
In the KVBC dedup, the way that defaults were set for some replica
parameters changed. This led to our default config's specification of
`concord-bft_size_of_reserved_page = 0` causing a divide-by-zero in
replica initialization here:
https://github.com/vmware/concord-bft/blob/master/bftengine/src/bftengine/ClientsManager.cpp#L44.

Regeneration produces a larger diff than strictly necessary to solve
this issue, but it has been a while since we updated these, so let's
bite the bullet.