ThinReplica: Refactor & sync'ed subscribe
The subscribe interface needs to do similar tasks as the KvbAppFilter, hence
let's refactor and rename the underlying KvbFilter functions. In addition, the
two subscription APIs are now able to deliver "old" blocks as well. The user
specifies a block id and with that the subscription APIs read from KVB until
caught up with the last created block. Thereafter, we get new blocks/updates
from the commands handler directly.