Add the test commands handler and gRPC service
This adds the ability to run and test Concord without an execution
engine.  The vision for how this can be used for testing:

A test case is packaged into a gRPC request and sent to a Concord
node.  The TeeCommandsHandler will create a response per the
information which was passed in.  Further details are in the DAML
Privacy Testing test plan:

https://confluence.eng.vmware.com/display/BLOC/DAML+Privacy+Testing

This could also be used for other things.  For example, performance
of Concord can be tested independently, without having to think about
whether performance bottlenecks are due to Concord or the execution
engine being used.

Usage:
------
Generate a concord configuration.  In the concord config file, ensure
the following properties:
  daml_enable: false
  eth_enable: false
  tee_enable: true
  tee_service_addr: 0.0.0.0:50051 (unique per node if running natively)
Start concord.
Send a request using grpcurl:
  grpcurl -plaintext -d '{ "test_input" : "x" }' 127.0.0.1:50051 com.vmware.concord.tee.TeeService/RunTest
  {
    "test_output": "Test Execution Handler sees that you passed in 'x'"
  }