Add acknowledgement placeholder to TRC Library
This commit adds placeholder code to the Thin Replica Client Library for
the acknowledgement of updates in the
ThinReplicaClient::AcknowledgeBlockID function; the placeholder code
includes a basic implementation, but with the part to actually send the
acknowledgement commented out given we know acknowledgement is currently
unimplemented from the Thin Replica Server side.

The motivation for using this placeholder implementation rather than
having the unimplemented AcknowledgeBlockID function automatically fail
or print warnings every time it is called is to encourage Thin Replica
Client applications to be written correctly using acknowledgement, even
before acknowledgement is implemented on the Thin Replica Server side,
so as to minimize the changes that will be needed to the application
code once we have a server-side acknowledgement implementation.

This commit additionally uncomments the AcknowledgeBlockID calls from
the Thin Replica Client example application, revises
thin-replica-client/README.md to document the new expected output, and
revises the interface-documenting comments in the Thin Replica Client
Library header file to clarify the requirement for and usage of
acknowledgement.