ThinReplicaServer: Filter by TRIDs
This change introduces a new protobuf message `ValueWithTrids` in order to
attach Thin Replica IDs (TRIDs) to a key-value pair. When a Thin Replica client
subscribes or reads from a Thin Replica server it has to provide its own TRID
which is then used to filter key-value pairs. At the moment, this feature is
only used within the DAML execution stack and the list of TRIDs is always
empty. In a later change, the DAML execution engine (validator) will not only
return key-value pairs that need to be persistet but also a list of TRIDs for
each of them. As soons as this is done we can do proper end-to-end testing.