Initial support for pre-execution
This MR introduces the first in-product changes related to the new pre-execution workflow, support for which is also being added in Concord-BFT. The following changes are introduced here:
1) Enhance the Concord API with messages related to pre-execution.
2) Add support for pre-execution flags in command handlers' Execute() methods
3) Implement pre-execution in the case of DAML.

In case the "preExecute" flag is on, we simply need the execution engine to run the command, returning us the read & write sets, which we then save in the PreExecutionResponse object.No changes to the blockchain state are made - we simply return the PreExecutionResponse (as part of ConcordResponse). This goes back to Concord-BFT where it is subject to a newly introduced "pre-execution consensus" (f+1 votes).

Once this succeeds, the PreExecutionResponse is queued for BFT consensus and eventually comes back to the execution engine (with the 'hasPreExecuted' flag set to true). This part isn't implemented yet, but it will simply consist in checking the provided PreExecutionResponse for conflicts and if none, appending a new block with the pre-executed write_set.

For further details, you may refer to the high-level design document for Pre-execution:
https://confluence.eng.vmware.com/display/BLOC/Pre-execution+HLD

Note: the HLF and EVM execution engines do not provide direct access to the read set, which is required for pre-execution. Because of this, I suggest to focus on DAML for the time being.