Add unsubscription placeholder to the TRC Library
This commit adds placeholder code to the Thin Replica Client Library
for the ThinReplicaClient::Unsubscribe function. The Unsubscribe RPC
call is not yet implemented from the Thin Replica Server side; this
placeholder is intended to allow applications to use the unsubscribe
library function as they normally would until the server-side
functionality is available so that applications will not need to be
modified once the server-side functionality is available.

In addition to adding this placeholder, this commit also modifies
comments in the thin_replica_client.hpp header file in order to clarify
the role and usage of the ThinReplicaClient::Unsubscribe function,
modifies the Thin Replica Client example application to ensure it
demonstrates the Unsubscribe function, and revises
thin-replica-client/README.md to reflect the new behavior expected of
the example application.