Add failure condition registration to TRC Library
This commit adds support to the Thin Replica Client Library for
registerng and using failure conditions with a ThinReplicaClient object.
Specifically, this commit implements
ThinReplicaClient::RegisterSubscriptionFailureCondition and modifies the
ThinReplicaClient's subscription thread implementation such that the
thread can notify a registered failure conditioin if it reaches a state
where it can no longer make progress with an active subscription.

Additionally, this commit modifies the Thin Replica Client example
application to register a failure condition so it can report in its
output when the subscription fails, and it modifies
thin-replica-client/REAMDE.md to document the new behavior of the
example application.