Save a summarized time in every KVB layer block
Rationale is that this change will support getting the time at previous
blocks. A use case of it is time-based pruning (VB-2298).

Extend the TimeContract so that it allows saving/loading a summarized
time to/from a newly-introduced KVB key. Change
ConcordCommandsHandler::addBlock() so that it saves the summarized time
as returned by TimeContract::GetTime() at KVB block creation.

Additionally, provide a test for the newly-introduced TimeContract
functionality.