Adopt Marcin's Proposed Facade to the TRC Library.
This commit makes changes to the Thin Replica Client Library originally proposed and written by (@mziolek) in relation to his consumption of this library in DAML applications. The most notable of these changes is the introduction of a facade to the existing Thin Replica Client interface to simplify it and avoid its transitive include dependencies.

This commit also includes some minor revisions to Marcin's original changes.