Implement TRC Library Subscribe call with Block ID
This commit implements the previously unimplemented 2-parameter overload of
ThinReplicaClient::Subscribe; the second parameter this overload accepts that
the existing implementation doesn't is a Block ID, which can be used to resume
an existing from the latest known Block ID. This implementation is relatively
simple, as conveniently, the bulk of the logic that needs to be shared between
the two Subscribe overloads was already separated out into the function run by
the asynchronous thread the Subscribe calls launch.

This commit additionally modifies the Thin Replica Client example application in
order to have it demonstrate the overloaded call by stopping and resuming its
subscription, destroying and reconstructing its Thin Replica Client object in
between. thin-replica-client/README.md is also modified to reflect the new
expected behavior.