Initial changes for introducing pkvutils
Adds the dependency to the pkvutils library from the oem-kv-integration-kit.
Modifies the validator and participant-state implementations to use the pkvutils
fragmentation/defragmentation.

This is an initial first step towards introducing the privacy features of DAML.
The implementation will be later on cleaned up by moving most of the logic
in KVBCParticipantState, KVBCLedgerServer and KVBCValidator into wrapper
libraries.