KvbAppFilter: Remove sleep
Remove random sleep. The sleep was an overly cautious implementation in case
the consumer (the thin replica server) is too slow. In that case we don't want
to hog the CPU with the KvbAppFilter thread. However, we don't have and
data/evidence that this is the case hence we should remove it and introduce
delays (or an event driven approach) if performance measurements suggest that.