Change code related to "master runs" to be generic.
We have decided to have release branches for the time being instead
of staying on a master branch for everything.  This change replaces
instances of "master" with either the string "ToT" (in comments)
or a variable holding the branch name.

This change also disables the RegressionTests suite because the
Concord exit code has been non-zero for two test runs.  That is
not critical and not related to this change, and this change
is blocking people.