ThinReplicaClient: Increase max inbound msg size
The default size for incoming messages for gRPC clients is 4MiB. The DAR file
required to run the CHESS+ application is larger than that. We have encountered
the exact same problem for v0 and had to increase Concord's message size limit.
The message size limit is 16MiB, the DAR is less than 16MiB, and hence we use
that number as the limit for the thin replica client.

Close VB-2441