Add wait for server responsiveness to TRC library
This commit adds logic to the
thin_replica_client::ThinReplicaClientFacade constructor to have it wait
for up to a fixed amount of time after it constructs its gRPC channels
for the servers on the other end of those channels to become responsive
before proceeding to construct its ThinReplicaClient.

This may prevent the ThinReplicaClient from crashing or producing
exceptions as a result of trying to connect to servers while they are
still in a transient starting-up state; we speculate launching a Thin
Replica Client application concurrently or in immediate succession to
the cluster it is supposed to connect to may be a not-exceedingly-rare
occurence in testing scenarios.