Use version subdirectories in Artifactory
Aspects of our processes rely on getting the "latest" of things from
artifactory.  Now that we are doing development work on concurrent
branches, the latest will be whatever branch ran last, which will
lead to problems as branches diverge.

This change places a build within version subdirectory which
matches its branch. The subdirectory name is created from the same
file used for creating the build number.