Validating and committing a pre-executed request
This MR wires the second part of pre-execution: validation and commitment of the pre-execution result.
After the initial wiring MR, the ConcordCommandsHandler receives the result of a pre-executed request as a ConcordRequest (request_) object containing a PreExecutionResult.

The "entry point" of the changes presented here is located at concord_commands_handler.cpp:128 where we enhance the current execution logic with support for a pre-executed request. All logic related to normal execution, pre-execution or committing pre-execution results is implemented for each engine in the virtual Execute() method.
In this MR we focus on the DAML implementation.

DAML's Execute() starts with a check if the provided request contains a pre-execution result. If so, it checks it for conflicts and records it. Although not specific for DAML, the HasPreExecutionConflicts() check needs to happen here, because in the future, DAML will have a special write_set for when a conflict has occurred.

If no conflicts were detected, commit the transaction using the newly factored-out recordTransaction() method. The normal execution workflow is unchanged and at the end also calls the recordTransaction() method.

PS: we need the original_request_id in the PreExecutionResult, in order to keep the correlation between the request and the produced block, if successful & valid pre-execution. This is why we have introduced the GetCorrelationId() function which parses the DAML command to extract its correlation id.