[VB-2165]: Add support for an external client library with configuration
The external client library provides a Concord BFT client. Its purpose
is to provide a simple and generic interface and support configuration
via a file so that users don't need to know the configuration structure
(allowing for ease of use and changes without affecting users).

A new external client file format has been introduced. Sample files are
provided for both UDP and TLS connections.

ConcordConfiguration has been slightly refactored to expose
functionality that can be reused in other parts of the system.

The communication.hpp file has been moved to the config directory to
allow reuse from other parts of the system.

Tests are provided to verify the newly-added functionality.

NOTE: Client objects are designed to be safe in terms of memory
management (e.g. creating an object and it going out of scope should be
safe). However, there are a number of memory leaks in different
components that need to be fixed for that to hold. For example -
ClientImp leaks the bftClient_, TlsTCPCommunication leaks a thread
(_pIoThread) and so on. These issues will be handled in separate
commits.