[VB-2298]: Add support for time-based pruning
Extend the system's capability by supporting time-based pruning. To
achieve it, the pruning_duration_to_keep_minutes system-wide
configuration option is added. It specifies a time range (in minutes)
from now to the past that determines which blocks to keep and which are
older than (now - pruning_duration_to_keep_minutes) and can, therefore,
be pruned. If not specified, a value of 0 is assumed. If the
pruning_num_blocks_to_keep option is specified too, the more
conservative pruning range of the two will be used (the one that prunes
less blocks). The pruning_duration_to_keep option requires the time
service to be enabled.

In order to implement, the KVBPruningSM expects that the summarized time
generated by TimeContract is written to KVB storage by
ConcordCommandsHandler. A binary search on the non-decreasing block time
is performed across the blockchain via the BlockchainView class from
concord-bft. A TimestampedBlockInfo class is added to support
BlockchainView's requirements and ensuring compatibility with standard
algorithms (namely std::lower_bound).

A third pruning option is added - pruning_enabled . It is a system-wide
configuration option that indicates if pruning is enabled for the
replica. If set to false, LatestPrunableBlockRequest will return 0 as a
latest block(indicating no blocks can be pruned) and PruneRequest will
return an error. If not specified, a value of false is assumed.

The KVBPruningSM tests have been updated to show the added functionality
and the dependencies with existing functionality work as expected.

NOTE: Minor changes related to pre-execution/pre-processing are required
to support the latest version of concord-bft .