Fix Race when Ending Subscriptions in TRC Library.
This commit is intended to address assertion-failure-triggering race
conditions between the Thin Replica Client subscription thread and
functions recalling the subscription thread; this is intended to fix the
bug tracked by VB-2522. This commit also includes a few other minor
related fixes to ensure the Thin Replica Client Library should smoothly
handle stopping, ending, resuming, and starting new subscriptions.

Fixes included in this commit:
- An incorrect assertion is removed from the Thin Replica Client
subscription thread function (ThinReplicaClient::ReceiveUpdates); the
assertion asserted that the thread had not been recalled when it first
starts running, but this may not be the case as in some circumstances
the thread's first running can race with its recall. The assertion is
replaced with a condition to just have the thread finish immediately if
it sees it was recalled before it could run.
- A few additional lines are added towards the end of
ThinReplicaClient::ReceiveUpdates() are added to ensure its cleanup is
comprehensive so it will come into its next run clean if a new
subscription is opened.
- Logic is add to both ThinReplicaClient::Subscribe overloads to ensure
they properly stop any existing subscription before creating a new one.