Add Data Subscription Failover to TRC Library
This commit enhances The Thin Replica Client Library’s implementations
of ThinReplicaClient::Subscribe and ThinReplicaClient::ReceiveUpdates
(the latter of these being the thread that asynchronously drives a
ThinReplicaClient’s active subscription) in order to allow the
ThinReplicaClient to fail over its data subscription to another server
if a data stream read fails, or if it finds the data stream to be in
disagreement with the hash streams.

This commit both modifies ThinReplicaClient::ReceiveUpdates to support
failover of the data stream in an active subscription and
ThinReplicaClient::Subscribe to support trying different servers for
initial state data if it cannot find servers in agreement with the
state it pulls from the first server.