ThinReplica: Correct block id documentation
From the initial design, we used to have a "last known cursor" which eventually
became "last known block id". However, the latter doesn't apply anymore due to
the application can compute the next block ID by itself. The cursor was opaque
and hence this wasn't possible. With this change we remove the confusion of the
last known block id.
While changing the parameter name in daml, I re-ran the code format script
which is why there are format changes as well.
BC-361