Fix dangerous memory management
The former code was built on top of the assumption that the inner buffer
of std::string is always allocated in the heap.
However, due to the Small String Optimization, the buffer may be
allocated on stack in the case of a small string.
Thus, the deletion of the inner buffer from sliver may lead to a
segmentation fault.