[Pre-execution] Checking the write set for conflicts
In addition to the read set, we also need to check that the write set doesn't contain conflicts. This also covers the scenario mentioned by Jussi:

A situation where this might happen is when a key does not already exist and two requests try to read and update this key. Say A and B are clients and k is a specific key and we run the following pre-executions in parallel:
1. A reads k, it does not exist and writes k=1
2. B reads k, it does not exist and writes k=10

Pre-execution results are sent to be committed, 1. comes first, 2. second.

Now we need to make sure that request 2. fails. I'm not sure if this is already covered (e.g. by making sure key 'k' is included in the read-set, even if it didn't exist), but just wanted to point this out.

PS: this MR includes two unit tests
* conflicting_write_set: a normal case of a stale write
* conflicting_write_of_new_key: the case described by Jussi above