Refactor the DAML ExecuteCommit workflow
The purpose of this MR is to clean-up the implementation of the ExecuteCommit() method, which had become spaghetti code and hard to read (hence hard to reason about).
Non-Regression is ensured by:
- DamlTest unit test
- Hermes DAML test suite
- cleaner code and easier review/correctness proof

After this refactoring, the ExecuteCommit method should only define the main workflow of executing a DAML commit request. All the "elementary" steps are now defined in separate private methods.