[Minor] Pre-execution commit receives a ConcordResponse, not a ConcordRequest
This MR only impacts the DAML commands handler unit test, in the part where it tests the commit of an already pre-executed request (has_pre_executed = true).

In our pre-execution flow, the PreExecutionResult arrives wrapped in a ConcordResponse, instead of a ConcordRequest. It is then deserialized and wrapped in a ConcordRequest, allowing it to seamlessly integrate with the rest of the command handler's workflow.

We want the DAML unit test to be as close as possible to the actual pre-execution flow, hence the changes proposed in this MR.

PS: the unit test also works if the PreExecutionResult is wrapped in a ConcordRequest. This is because ConcordRequest and ConcordResponse look the same on the Protobuf wire.