Running SKVBC via Test Execution Engine
This MR is the first of a series of steps in an effort to execute Apollo BFT tests against production-like Concord deployments.
A detailed design document is available here:
https://confluence.eng.vmware.com/display/BLOC/Running+Apollo+in+product+CI+via+Hermes

The goal here is to wire the Simple K/V Blockchain (SKVBC) execution engine into Concord.

The entry point is via the TEE gRPC endpoint, where we send either SkvbcRead or SkvbcWrite with a RawSkvbcRequest containing the serialized messages from the SKVBC protocol (packed C-structs). The RawSkvbcRequest is mapped to a ConcordResponse with a TeeRequest + SkvbcRequest and passed on to the BFT client pool for consensus & execution. The response path is built symmetrically.

The SKVBC execution logic is implemented by the SimpleKVBC/InternalCommandsHandler. The TeeCommandsHandler essentially acts as an "adapter" between the "TEE" and the SKVBC protocol.

Note: I have introduced a request_context_ (ConcordRequestContext), so that the TeeCommandsHandler can provide the InternalCommandsHandler with all required "protocol context" data (including client ID, sequence number, max reply size).