persephone to filter out metrics docker images
Currently, there is no way to limit the size of the buffer written by
wavefront when it is not able to reach the url. In our usecase, most
number of times, we will not have a url set up. Wavefront team is
working on a feature which will not allow wavefront to come up for
incorrect of null values of url and token. This is a temporary fix to
not deploy the containers when user has not provided an URL or token.