Bootstrap Apollo BFT pytest suite
This MR adds a pytest suite which bootstraps Apollo and runs a single basic test (test_get_block_data).

To do so, we need to import existing Apollo Python modules. This proved somewhat tricky, due to having a 'util' module in both Apollo and Hermes. I haven't so far found a better way to tackle this than to tweak sys.modules...

Once Apollo modules are loaded, we need to create a representation of the containerized blockchain as a 'BftTestNetwork' instance that Apollo understands. This is done in the `bft_network` fixture calling the BftTestNetwork.existing() classmethod factory.

Three essential elements are required for Apollo tests to run against a VMware Blockchain deployment:
- commands for starting and stopping replicas (in this case, using docker)
- a description of the replicas as `Replica` objects
- a set of initialized clients (ConcordSkvbcClient) - which know how to talk to the SKVBC application

The `bft_network` fixture is then injected into test_skvbc_get_block_data, which uses Trio to invoke the underlying Apollo implementation of the test.