Make TEE requests non-readonly by default
## The problem
From the BFT perspective, read-only requests are treated differently than transactions that modify the state.
In the read-only case, replicas do NOT go through consensus. Instead, for read-only requests, each replica simply sends its current state.
The BFT client then waits for 2f+1 matching replies and returns.

For non-readonly requests, on the other hand, the full BFT consensus kicks-in (request ordering, followed by sequential, deterministic execution).
Again, the BFT client waits for 2f+1 matching replies and returns.

The TEE's RunTest endpoint currently treats every request as read-only, which only activates the read path of the BFT consensus.

## Suggested solution (short-term)
In this MR, I suggest to set the `isReadOnly` flag to `false`, by default. This is OK, because the only current use case, PrivacySanityTest, is an actual transaction that modifies the state (and therefore needs to go through the full BFT consensus). This is the actual state transition of the replicated state machine.

## Suggested solution (long-term)
Going forward, if we need to use the TEE endpoint for read-only requests too, we should enhance the `TeeRequest` message with an explicit `isReadOnly` flag.