BC-1778: Trace context from TR Server to TR Client
This commit is part of the span starting after a block is added to DB
and ending in the Indexer at the DAML side.
The commit covers the next part of the flow:
* Creates a child span("trclient_block_verification") from span context of
the incoming block.
* Creates "trclient_verify_hash" to track the block's hash verification.
* In case additional servers are needed to verify the block, creates
"trclient_verify_hash_against_additional_servers".
* In case of additional block verification, creates a span
"trclient_verify_hash_against_additional_servers".
* Injects the first span into the outgoing update as a byte array.

Additional changes:
* Some refactoring was made for the open tracing's sake.
* Docker file was updated to add opentracing and jaeger libraries.
* Jeager agent was added to docker/docker-compose-daml-metrics.yml
* Added jaeger_agent field to the JNI integration layer