ThinReplicaTests: Attempt to fix intermittent failures
Within the test we query the Block ID from _one_ replica and use it to query
all replicas. However, given that not all replicas might be in the same state
the Block ID might not be available for some replicas. Therefore, instead,
query all replicas for their newest Block ID and choose the smallest. Thin will
increase the test time the more replicas we deploy but should be negligible for
now.