Improve Racetrack reporting structure
- Instead of `set` being a test suite, make the set be the Jenkins run,
on which testSuite/testCases are reported on

- Removed Racetrack publish from `event_recorder.py` since initiating
set would only be in the beginning and the end of the run

- Modified groovy file, set begin by `invoke.py racetrackSetBegin` and
end by `invoke.py racetrackSetEnd`

- Modified `getTestingEnvironmentInfo` function in `racetrack.py` to
conform this new structure. Four branches: {MR, master, release, other}

- Modified `caseStart` in `racetrack.py` such that calling this function
no longer requires `setId`, this will be understood to be whatever set in
the beginning of the Jenkins run. The test case reporting should now focus
more on `suiteName`, `caseName`, and `description` of the case.

- Modified `helper.py` to dictate where Racetrack Set ID will be stored for
the duration of the run (/blockchain/vars/racetrack_set_id.json)