Add a mechanism to only build what is needed for a change.
As with the change to define tests to run for a change, a directory
can contain this file:

        Name: tests_and_artifacts.json
        Format:
        {
          "test_suites": [],
          "build_targets": []
        }

build_targets is a list of the docker images to build, such as "concord-core".
Anything not listed there will be pulled from Artifactory instead.

Also fixes a typo in the regex in make-prebuilt-env.sh.