TRC: Refactor update loop
This change attempts to simplify the update loop. In short, we start with the
data stream at replica 0 and try to read data. No matter whether we are
successful, thereafter we try to find a hash agreement amongst all servers.
Thereby, this change consolidates the existing two hash-verify functions into
one and sorts the list of server indexes beforehand so that open streams are
queried first. Thereafter, we check that the data we have (or not have) is part
of the agreeing servers. If not, then we close the exisitng data stream and try
to get the data from one of the agreeing servers instead.
In addition, we separate out one data and one hash stream creation function and
also replace most nested if-else statements with early return/continue/break
statements. The latter results in more verbose/lines of code but, IMHO, makes
it much easier to read and understand.