TRC: Timeout Read() for data and hash streams
So far, we rely on all replicas to respond. However, replicas can become too
busy and/or might not be ready to respond to TRC requests or simply run into
a live-lock. One example is state transfer. While in ST, the replica doesn't
participate in conensus and thereby the Concord commands handler doesn't
forward new updates (or any update) to the TRS. With this change, we expect the
replica to respond within a given amount of time. Note, gRPC Read() doesn't
provide the ability to set timeouts but only set deadlines (absolute
timestamps) per stream context. In order to keep our stream alive and still
time out for long running Read() calls, we wait on hash and data stream reads
asynchronously in a thread per call. Given that the TRC is single threaded
there is always only one thread waiting for a Read() call.
With this commit the update loop changes. We wait indefinitely until we find a
replica that returns valid data. Thereafter, we still try to get the usual
agreement but we are less patient. Given that one replica has returned data we
expect that all other replicas should have the data too and thereby are able to
respond with a hash in time.
If a Read() request times out then we send TryCancel() to the server and close
the connection. On the server side, we keep checking the cancel flag and
terminate the subscription as soon as we see it.