vDAML: Sync log formats
The ledger API and the execution engine use different log formats than the thin
replica client and Concord. This change closes the gap so that all log messages
start with an "almost" ISO8601 timestamp. This will make log aggregation with
Concord much easier, especially when you use a program such as `lnav`.
Note: We don't print the thread id for info logs. For the TRC, there is only
one thread and the DAML components print very long thread id strings. However,
the tread ids are recorded in the log file.