Python client for Concord DAML gRPC endpoint
The ConcordDamlClient class allows us to talk directly to the DAML gRPC endpoint on Concord nodes (without passing through the Ledger API). This allows us more flexibility when testing the DAML flavor of Concord. It is especially useful when we want to pass specific execution flags, such as read-only (0x1), pre-execution (0x2), or write (0x0 - the default one).

== Note to the reviewers ==
This class will be used in the upcoming Hermes tests for DAML pre-execution.
Ideally, we should send the requests via Ledger API, to avoid for example depending on the Concord gRPC API for DAML, which will be deprecated with the introduction of the external BFT client. However, neither the pre-execution flag, nor the external BFT client are yet integrated in the Ledger API Scala implementation.

This is why I suggest the ConcordDamlClient as a workaround, in order to enable pre-execution system testing with DAML in the loop.